Refactored the exejobinfo - added a start method which is invoked from the runner. The reason for this was to support batch jobs not only using LSFJob but also with drmaa jobs or even local process runner
Former-commit-id: c49ac01c19c1ba16a6b517a6cd48182a289b411e

+++ jacs2-services/src/main/java/org/janelia/jacs2/asyncservice/common/LocalExeJobInfo.java
@@ -2,17 +2,20 @@ package org.janelia.jacs2.asyncservice.common;
 
 import org.janelia.model.service.JacsJobInstanceInfo;
 
+import java.io.IOException;
 import java.util.Collection;
 import java.util.Collections;
 
 public class LocalExeJobInfo implements ExeJobInfo {
-    private final Process localProcess;
+    private final ProcessBuilder localProcessBuilder;
     private final String scriptName;
-    private boolean done;
-    private boolean failed;
+    private Process localProcess;
+    private volatile boolean done;
+    private volatile boolean failed;
+    private volatile boolean terminated;
 
-    public LocalExeJobInfo(Process localProcess, String scriptName) {
-        this.localProcess = localProcess;
+    public LocalExeJobInfo(ProcessBuilder localProcessBuilder, String scriptName) {
+        this.localProcessBuilder = localProcessBuilder;
         this.scriptName = scriptName;
     }
 
@@ -22,11 +25,29 @@ public class LocalExeJobInfo implements ExeJobInfo {
     }
 
     @Override
+    public String start() {
+        if (!terminated) {
+            try {
+                localProcess = localProcessBuilder.start();
+                return localProcess.toString();
+            } catch (IOException e) {
+                done = true;
+                failed = true;
+                throw new IllegalStateException(e);
+            }
+        } else {
+            return null;
+        }
+    }
+
+    @Override
     public boolean isDone() {
         if (done) return done;
-        done = !localProcess.isAlive();
-        if (done) {
-            failed = localProcess.exitValue() != 0;
+        if (localProcess != null) {
+            done = !localProcess.isAlive();
+            if (done) {
+                failed = localProcess.exitValue() != 0;
+            }
         }
         return done;
     }
@@ -43,7 +64,13 @@ public class LocalExeJobInfo implements ExeJobInfo {
 
     @Override
     public void terminate() {
-        localProcess.destroyForcibly();
+        terminated = true;
+        if (localProcess != null) {
+            localProcess.destroyForcibly();
+        } else {
+            done = true;
+            failed = true; // consider an error if early terminated;
+        }
     }
 
 }
